dom/media/tests/mochitest/test_dataChannel_noOffer.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="pc.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "856319",
    title: "Don't offer m=application unless createDataChannel is called first"
  });

  runNetworkTest(function () {
    var pc = new mozRTCPeerConnection();

    // necessary to circumvent bug 864109
    var options = { offerToReceiveAudio: true };

    pc.createOffer(function (offer) {
      ok(!offer.sdp.contains("m=application"),
        "m=application is not contained in the SDP");

      networkTestFinished();
    }, generateErrorCallback(), options);
  });

</script>
</pre>
</body>
</html>
